chore: Revamp Metadata on top of PD Client Watcher (#981)
* Remove fetcher and prepare to replace it with pd-client


* Simplify metadata cache in range server


* Define MetadataObserver trait


* Implement RangeManager on top of MetadataManager using observer pattern


* WIP


* WIP


* Fix format


* Await range manager metadata loading


* Aggregate store in RangeManager


* RangeManage aggregates ObjectStorage for now


* Remove Store from Worker/Session/ServerCall and various handlers, delegate ops to RangeManager instead


* Get ready to fix tests


* Fix handler::append::tests::test_apply


* Fix unit test handler::create_range::tests::test_create_range_apply


* Fix test_create_range_apply_when_store_fails


* Fix tests


* Make tests pass


* Fix fmt issue


* Fix clippy warnings


* Use ulog


---------

+++ components/model/src/request/fetch.rs
@@ -24,7 +24,7 @@ impl From<&FetchRequest> for FetchRequestT {
         let mut res = FetchRequestT::default();
         res.max_wait_ms = value.max_wait.as_millis() as i32;
         let mut range = RangeT::default();
-        range.stream_id = value.range.stream_id();
+        range.stream_id = value.range.stream_id() as i64;
         range.index = value.range.index();
         range.start = value.range.start() as i64;
         if let Some(end) = value.range.end() {
